Add files via upload
This file contains the prototype Matt uploaded to the google drive. I edited it to read and display the first line of a .txt file. With these edits, when the player runs out of lives, the question from the .txt file is displayed on the screen. To work on next could be: allowing the player to enter an answer to the question and recording that answer (back to a .txt file?), and allowing the game to continue after the question has been answered. I put my name next to the code I edited and added on the .py file.